Use singleton to fix 🩹o3ds multiple instance 🐞
o3d's CPP pybind implementation currently causes multiple constructions (2) during binding from python o3d down to the CPP in the pipelines.
This fixes by implementing a singleton for TransformationEstimation objects.
Parallelism is implemented lower in the cpp objects and since python GIL is a hard limit , this *should* not cause a limitation.